chiisana_tenohira_mad: Initialize
This commit is almost about the initialization of this project. It adds
a README.md in the project folder and adds some subtitles and some
content about the product. The project README should be inherited from
a template but there hasn't any of it so that the version string is
blank temporarily.