chiisana_tenohira_mad_2: Complete README and enhance folder
This commit contains some changes:

**Add:**

- `README.md` in creation folder has been completed. It shouldn't be
modified unless the post is refused.
- The ASS subtitle has been put into `src` folder.

**Remove:**

- README files in the repository root folder have been deleted. They
will cause a merge failure due to the different contents of the files in
this branch and `main` branch.

Note that the structure of `README.md` has some small changes which are
different from one of `README.md` in `chiisana_tenohira_mad`. This file
will become a new standard.