audio_visualization: Complete generation and add test
`generate_video.py` and other scripts have been improved and it can
generate videos successfully. And there are more changes for testing
this project.

Note that the code still needs more improvement because OpenCV may
create a big video file, which we almost don't need that. This will be
fixed by FFmpeg.