audio_visualization: Add fade-in and more stuff
This commit has many changes. Key changes are as below:

**Add:**
- Add fade-in and fade-out (just keep silent until the whole wave fades
out) to the creation.

**Modify:**
- The original generation function has been split into three parts due
to the additional special effects of the creation.
- `VideoConductor` has had a new initialization function, which contains
more parameters to support videos with various needs. Visit FFmpeg
official site for more information about them.
- Some parameters of `VideoConductor.__init__()` can be `None` now.

Note that this commit is not an end of this creation. We still need more
special effects.